Pull request correction
Improved the code, regarding the feedback of the pull request
